+++ sdk/keyvault/azkeys/utils_test.go
@@ -169,7 +169,7 @@ func startTest(t *testing.T, MHSMtest bool) *azkeys.Client {
 	if MHSMtest {
 		URL = mhsmURL
 	}
-	opts := &azkeys.Options{ClientOptions: azcore.ClientOptions{Transport: transport}}
+	opts := &azkeys.ClientOptions{ClientOptions: azcore.ClientOptions{Transport: transport}}
 	return azkeys.NewClient(URL, credential, opts)
 }